added product, review, and user repository public interfaces (accessible by the package)
Jaccard Distance: a logged-in user should be able to view the Jaccard distance between themselves and other users.
Degree of Separation: a logged-in user should be able to view the degree of separation between themselves and other users.
User Searching 1: allows a logged-in user to find friends (other users) by viewing a list of users sorted by the Jaccard distance between themselves and the user.
User Searching 2: allows a logged-in user to find friends (other users) by viewing a list of most-followed users.